requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.95.0
fonttools==4.39.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jedi==0.18.2
kiwisolver==1.4.4
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
stack-data==0.6.2
starlette==0.26.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
